requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary>=2.9.11
alembic==1.13.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
pydantic>=2.12,<3
pydantic-settings>=2.8,<3
email-validator==2.2.0
celery==5.4.0
redis==5.0.8
boto3==1.35.25
botocore==1.35.25
stripe==10.12.0
openai==1.46.0
httpx==0.27.2
pypdf==5.0.1
pdf2image==1.17.0
python-docx==1.1.2
Pillow>=11.3.0
pytesseract==0.3.13
pytest==8.3.3
pyotp==2.9.0
clamd==1.0.2
itsdangerous==2.2.0
prometheus-client==0.21.0
openpyxl==3.1.5
xlrd==2.0.1